Adding fax4 (T6) compression
